splitting source code into different polyfills for better reusability
