+++ lib/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Loader: require('./loader.js'),
+  Module: require('./module.js'),
+  System: require('./system.js')
+};
